Actions and Named Routes
Express's default router/handlers do not have a way of working with
named or reversible routes which means creating urls by hand which is
fragile, prone to errors and hard to refactor.

We are also creating routes in various places within the application
instead of a single central place which is hard to reason about.

To address these issues we have created a router middleware that allows
us to structure our application as:

* Actions (handler-like functions that do some work and return JSON data)
* Routes (mapping a url to/from an Action)
* Components/Templates (macros/templates that render the data returned
  from Actions)