Single user mode: skipping oauth for local dev
paas-admin requires wiring up to the UAA server to perform the oauth
dance which requires configuring a UAA client. This makes local develop
against an existing cloudfoundry deployment fiddly since you need to
modify the UAA client to point it's redirect_uri at localhost.

Since most of the time we don't care about the oauth flow, we just want
to be logged in, we could skip the oauth authentication entirely and
pass an oatuh token for our user into the environment effectively making
it a single-user version of paas-admin.

This is a proof of this concept that allows setting
SINGLE_USER_ACCESS_TOKEN to an oauth token to skip auth which
drastically cuts down the amount of environment variable boilerplate
required to boot the app:

SINGLE_USER_ACCESS_TOKEN=$(cf oauth-token) \
API_URL="https://api.cloud.service.gov.uk" \
  npm start